+++ src/main/kotlin/us/appfluent/xwidget/actions/CommandAction.kt
@@ -4,9 +4,9 @@ import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.ToolWindowManager
-import org.jetbrains.plugins.terminal.ShellTerminalWidget
 import org.jetbrains.plugins.terminal.TerminalToolWindowFactory
 import org.jetbrains.plugins.terminal.TerminalToolWindowManager
+
 
 abstract class CommandAction(private val cmd: String) : AnAction() {
     companion object {
@@ -19,16 +19,24 @@ abstract class CommandAction(private val cmd: String) : AnAction() {
     }
 
     private fun runCommand(project: Project, cmd: String) {
-        val terminalManager = TerminalToolWindowManager.getInstance(project)
-        val window = ToolWindowManager.getInstance(project).getToolWindow(TerminalToolWindowFactory.TOOL_WINDOW_ID)
-        val contentManager = window?.contentManager
-        val widget = when (val content = contentManager?.findContent(TAB_NAME)) {
-            null -> terminalManager.createLocalShellWidget(project.basePath, TAB_NAME)
+        val toolWindow = ToolWindowManager.getInstance(project).getToolWindow(TerminalToolWindowFactory.TOOL_WINDOW_ID)
+        val contentManager = toolWindow?.contentManager
+        val terminal = when (val content = contentManager?.findContent(TAB_NAME)) {
+            null -> {
+                val terminalManager = TerminalToolWindowManager.getInstance(project)
+                terminalManager.createShellWidget(project.basePath, TAB_NAME, true, false)
+            }
             else -> {
-                contentManager.setSelectedContent(content)
-                TerminalToolWindowManager.getWidgetByContent(content) as ShellTerminalWidget
+                val selectRunnable = Runnable {
+                    contentManager.setSelectedContent(content, true)
+                }
+                when (toolWindow.isActive) {
+                    true -> selectRunnable.run()
+                    false -> toolWindow.activate(selectRunnable, true, true)
+                }
+                TerminalToolWindowManager.findWidgetByContent(content)
             }
         }
-        widget.executeCommand(cmd)
+        terminal?.sendCommandToExecute(cmd)
     }
 }