+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+### 0.1.2 (Oct 8, 2024)
+
+* Added context aware popup menu items to navigate to controllers and fragments. 
+* Changed project id from `us.appfluent.xwidget-jetbrains-plugin` to `us.appfluent.xwidget-intellij-plugin`
+* Updated Gradle Plugin from 1.x to 2.x
+* Removed uses of deprecated classes and methods
+
 ### 0.1.1 (Oct 5, 2024)
 
-* Added Tools menu options to initialize project, view documentation, and view issues. 
+* Added 'Tools' menu items to initialize project, view documentation, and view issues. 
 * Migrated to Gradle 2.x from 1.x
 
 ### 0.1.0 (Oct 4, 2024)